+++ src/utils/transactions/getTransactionValue/helpers/getTitleText.ts
@@ -1,7 +1,5 @@
-import { DECIMALS } from 'lib/sdkDappUtils';
+import { DECIMALS, formatAmount } from 'lib/sdkDappUtils';
 import { TokenArgumentType } from 'types/serverTransactions.types';
-import { formatAmount } from 'utils/operations/formatAmount';
-
 import { getIdentifierType } from 'utils/validation/getIdentifierType';
 import { getTransactionActionNftText } from '../../getTransactionActionNftText';
 import {